refactor(query): converge internal command-resolution imports on canonical seam
